refactor Users

+++ src/App/Users/UserRepositoryPdo.php
@@ -2,10 +2,11 @@
 
 declare(strict_types=1);
 
-namespace App\RegUsers;
+namespace App\Users;
 
-use PDO;
 use InvalidArgumentException;
+use PDO;
+use PDOException;
 
 final class UserRepositoryPdo implements UserRepositoryInterface
 {
@@ -13,6 +14,12 @@ final class UserRepositoryPdo implements UserRepositoryInterface
 
     public function __construct(private PDO $db) {}
 
+    /**
+     * {@inheritDoc}
+     *
+     * @throws InvalidArgumentException
+     * @throws PDOException
+     */
     public function find(string $key, $value): array|null
     {
         if (!in_array($key, $this->fields)) {
@@ -25,34 +32,50 @@ final class UserRepositoryPdo implements UserRepositoryInterface
         return $sth->fetch(PDO::FETCH_ASSOC);
     }
 
-    public function add(array &$user): bool
+    /**
+     * {@inheritDoc}
+     *
+     * @throws InvalidArgumentException
+     * @throws PDOException
+     */
+    public function add(array $user): array
     {
+        if (empty($user['username'])) {
+            throw new InvalidArgumentException('Username is required');
+        }
         $user = array_intersect_key($user, array_flip($this->fields));
+        $user['created_at'] = date('Y-m-d H:i:s');
+        $user['updated_at'] = date('Y-m-d H:i:s');
         $sql = "INSERT INTO users (username, password, created_at, updated_at)
                 VALUES (:username, :password, :created_at, :updated_at) RETURNING *";
         $sth = $this->db->prepare($sql);
-        $user['created_at'] = date('Y-m-d H:i:s');
-        $user['updated_at'] = date('Y-m-d H:i:s');
         $sth->execute($user);
-        $user = $sth->fetch(PDO::FETCH_ASSOC);
 
-        return $user;
+        return $sth->fetch(PDO::FETCH_ASSOC);
     }
 
-    public function update(array &$user): bool
+    /**
+     * {@inheritDoc}
+     *
+     * @throws InvalidArgumentException
+     * @throws PDOException
+     */
+    public function update(array $user): array
     {
+        if (empty($user['id'])) {
+            throw new InvalidArgumentException('User ID is required');
+        }
         $user = array_intersect_key($user, array_flip($this->fields));
+        $user['updated_at'] = date('Y-m-d H:i:s');
         $sql = "UPDATE users SET username = :username, password = :password, updated_at = :updated_at
                 WHERE id = :id RETURNING *";
         $sth = $this->db->prepare($sql);
-        $user['updated_at'] = date('Y-m-d H:i:s');
         $sth->execute($user);
-        $user = $sth->fetch(PDO::FETCH_ASSOC);
 
-        return $user;
+        return $sth->fetch(PDO::FETCH_ASSOC);
     }
 
-    public function count(array $filter = []): int
+    public function count($filter = []): int
     {
         $sql = "SELECT COUNT(*) FROM users";
         $sth = $this->db->prepare($sql);
@@ -61,9 +84,21 @@ final class UserRepositoryPdo implements UserRepositoryInterface
         return $sth->fetchColumn();
     }
 
-    public function filter(array $filter = [], $order = 'id', int $limit = 10, int $offset = 0): array
+    public function filter($filter = [], $order = '', int $limit = 10, int $offset = 0): array
     {
-        $sql = "SELECT * FROM users";
+        if (empty($order)) {
+            $order = 'id';
+        }
+        if (!in_array($order, $this->fields)) {
+            throw new InvalidArgumentException("Invalid key: {$order}");
+        }
+        $sql = "SELECT * FROM users ORDER BY {$order}";
+        if ($limit > 0) {
+            $sql .= " LIMIT {$limit}";
+        }
+        if ($offset > 0) {
+            $sql .= " OFFSET {$offset}";
+        }
         $sth = $this->db->prepare($sql);
         $sth->execute();
 
@@ -72,9 +107,12 @@ final class UserRepositoryPdo implements UserRepositoryInterface
 
     public function delete(array $user): bool
     {
+        if (!isset($user['id'])) {
+            throw new InvalidArgumentException('User ID is required');
+        }
         $sql = "DELETE FROM users WHERE id = :id";
         $sth = $this->db->prepare($sql);
-        $sth->execute($user);
+        $sth->execute(['id' => $user['id']]);
 
         return $sth->rowCount() === 1;
     }
